+++ shared/src/commonMain/kotlin/ui/viewmodels/authentication/PresentationViewModel.kt
@@ -12,6 +12,7 @@ import at.asitplus.signum.supreme.UserInitiatedCancellationReason
 import at.asitplus.valera.resources.Res
 import at.asitplus.valera.resources.warning_authentication_cancelled
 import at.asitplus.wallet.app.common.WalletMain
+import at.asitplus.wallet.app.common.extractCertificateChain
 import at.asitplus.wallet.lib.agent.SubjectCredentialStore
 import at.asitplus.wallet.lib.data.CredentialPresentation
 import at.asitplus.wallet.lib.data.CredentialPresentationRequest
@@ -50,18 +51,7 @@ class PresentationViewModel(
         descriptors = listOf() // TODO Replace with ISO Device Retrieval
         this.finishFunction = finishFunction
         this.sessionTranscript = sessionTranscript
-        this.verifierCertificateChain = (parsedRequest.readerAuthAll?.firstOrNull()
-            ?: parsedRequest.docRequests.firstNotNullOfOrNull { it.readerAuth })
-            ?.let { cose ->
-                cose.protectedHeader.certificateChain
-                    ?: cose.unprotectedHeader?.certificateChain
-            }
-            ?.let { rawChain ->
-                rawChain.map { bytes ->
-                    runCatching { X509Certificate.decodeFromDer(bytes) }.getOrNull()
-                        ?: return@let null
-                }
-            }
+        this.verifierCertificateChain = parsedRequest.extractCertificateChain()
     }
 
     override val transactionData = null